pyproject.toml
[project]
name = "mini-ai-chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[google-genai]>=0.3.26",
    "python-dotenv>=1.1.1",
    "pywebview>=5.4",
]
